Conditionally bootstrap rake

+++ deploy_in_container.sh
@@ -36,7 +36,14 @@ done
 
 cd $SCRIPTDIR
 rake -f deploy_in_container.rake align_ruby
-gem install rake
+
+# Bootstrap rake
+if [ -e /usr/local/bin/ruby ]; then
+  gem install rake
+else
+  gem uninstall rake
+fi
+
 which rake
 which ruby
 exec rake -f deploy_in_container.rake deploy_in_container
